ledgerize/pyproject.toml
[tool.poetry]
name = "ledgerize"
version = "0.1.0"
description = "CLI tool to normalize and report multi-bank statements"
authors = ["OpenAI <openai@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ledgerize", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pydantic = "^2.6.3"
pandas = "^2.2.1"
numpy = "^1.26.4"
python-dateutil = "^2.9.0.post0"
PyYAML = "^6.0.1"
rich = "^13.7.1"
plotly = "^5.20.0"
jinja2 = "^3.1.3"
sqlalchemy = "^2.0.27"
charset-normalizer = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
coverage = "^7.4.1"
ruff = "^0.1.11"
black = "^24.2.0"
mypy = "^1.8.0"
types-PyYAML = "^6.0.12.12"

[tool.poetry.scripts]
ledgerize = "ledgerize.cli:main"

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
